split AI chat remote wire tests
